# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Hope Apartments API",
    "version": "1.0.0",
    "description": "Internal and partner API documentation for the Hope Apartments service. Partner apartment access uses `X-API-Key` plus API key scopes. Internal routes accept Bearer JWT or the admin session cookie where documented. When using `/docs` in a logged-in admin browser session, Swagger requests can rely on that existing session automatically."
  },
...
  "paths": {
...
    "/api-keys": {
      "get": {
        "tags": [
          "API Keys"
        ],
        "summary": "List API keys",
        "description": "Lists partner API keys. Requires internal permission `api_keys:read`. Accessible to `admin` and `developer` with the current permission matrix.",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "adminSession": []
          }
        ],
        "x-required-permission": "api_keys:read",
        "responses": {
          "200": {
            "description": "API key list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid internal auth",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicError"
                }
              }
            }
          },
          "403": {
            "description": "Missing `api_keys:read` permission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicError"
                }
              }
            }
          },
          "503": {
            "description": "Auth or API key service is not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "API Keys"
        ],
        "summary": "Create API key",
        "description": "Creates a partner API key. Requires internal permission `api_keys:create`. With the current permission matrix this is effectively admin-only.",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "adminSession": []
          }
        ],
        "x-required-permission": "api_keys:create",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiKeyCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "API key created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyCreateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid payload or unsupported scopes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicError"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid internal auth",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicError"
                }
              }
            }
          },
          "403": {
            "description": "Missing `api_keys:create` permission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicError"
                }
              }
            }
          },
          "503": {
            "description": "Auth or API key service is not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicError"
                }
              }
            }
          }
        }
      }
    },
    "/api-keys/stats": {
      "get": {
        "tags": [
          "API Keys"
        ],
        "summary": "API key usage and lifecycle stats",
        "description": "Returns operational metrics for partner API keys. Requires internal permission `api_keys:read`. Accessible to `admin` and `developer`.",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "adminSession": []
          }
        ],
        "x-required-permission": "api_keys:read",
        "responses": {
          "200": {
            "description": "API key stats",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyStatsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid internal auth",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicError"
                }
              }
            }
          },
          "403": {
            "description": "Missing `api_keys:read` permission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicError"
                }
              }
            }
          },
          "503": {
            "description": "Auth or API key service is not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicError"
                }
              }
            }
          }
        }
      }
    },
    "/api-keys/{id}": {
      "get": {
        "tags": [
          "API Keys"
        ],
        "summary": "Read API key by public identifier",
        "description": "Reads a partner API key. Requires internal permission `api_keys:read`. Accessible to `admin` and `developer`.",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "adminSession": []
          }
        ],
        "x-required-permission": "api_keys:read",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "description": "API key public identifier (key prefix)."
          }
        ],
        "responses": {
          "200": {
            "description": "API key details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid API key identifier",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicError"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid internal auth",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicError"
                }
              }
            }
          },
          "403": {
            "description": "Missing `api_keys:read` permission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicError"
                }
              }
            }
          },
          "404": {
            "description": "API key not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicError"
                }
              }
            }
          },
          "503": {
            "description": "Auth or API key service is not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "API Keys"
        ],
        "summary": "Update API key",
        "description": "Updates a partner API key. Requires internal permission `api_keys:update`. With the current permission matrix this is effectively admin-only.",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "adminSession": []
          }
        ],
        "x-required-permission": "api_keys:update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "description": "API key public identifier (key prefix)."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiKeyUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "API key updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid path parameter, payload or unsupported scopes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicError"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid internal auth",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicError"
                }
              }
            }
          },
          "403": {
            "description": "Missing `api_keys:update` permission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicError"
                }
              }
            }
          },
          "404": {
            "description": "API key not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicError"
                }
              }
            }
          },
          "503": {
            "description": "Auth or API key service is not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicError"
                }
              }
            }
          }
        }
      }
    },
    "/api-keys/{id}/revoke": {
      "post": {
        "tags": [
          "API Keys"
        ],
        "summary": "Revoke API key",
        "description": "Revokes a partner API key. Requires internal permission `api_keys:revoke`. With the current permission matrix this is effectively admin-only.",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "adminSession": []
          }
        ],
        "x-required-permission": "api_keys:revoke",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "description": "API key public identifier (key prefix)."
          }
        ],
        "responses": {
          "200": {
            "description": "API key revoked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid API key identifier",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicError"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid internal auth",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicError"
                }
              }
            }
          },
          "403": {
            "description": "Missing `api_keys:revoke` permission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicError"
                }
              }
            }
          },
          "404": {
            "description": "API key not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicError"
                }
              }
            }
          },
          "503": {
            "description": "Auth or API key service is not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicError"
                }
              }
            }
          }
        }
      }
    },
    "/api-keys/{id}/reactivate": {
      "post": {
        "tags": [
          "API Keys"
        ],
        "summary": "Reactivate API key",
        "description": "Reactivates a partner API key. Requires internal permission `api_keys:update`. With the current permission matrix this is effectively admin-only.",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "adminSession": []
          }
        ],
        "x-required-permission": "api_keys:update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "description": "API key public identifier (key prefix)."
          }
        ],
        "responses": {
          "200": {
            "description": "API key reactivated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid API key identifier",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicError"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid internal auth",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicError"
                }
              }
            }
          },
          "403": {
            "description": "Missing `api_keys:update` permission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicError"
                }
              }
            }
          },
          "404": {
            "description": "API key not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicError"
                }
              }
            }
          },
          "503": {
            "description": "Auth or API key service is not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicError"
                }
              }
            }
          }
        }
      }
    },
    "/api-keys/{id}/rotate": {
      "post": {
        "tags": [
          "API Keys"
        ],
        "summary": "Rotate API key",
        "description": "Rotates a partner API key, returns a new secret once and revokes the previous key atomically. Requires internal permission `api_keys:rotate`. With the current permission matrix this is effectively admin-only.",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "adminSession": []
          }
        ],
        "x-required-permission": "api_keys:rotate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "description": "API key public identifier (key prefix)."
          }
        ],
        "responses": {
          "200": {
            "description": "API key rotated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyRotateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid API key identifier",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicError"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid internal auth",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicError"
                }
              }
            }
          },
          "403": {
            "description": "Missing `api_keys:rotate` permission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicError"
                }
              }
            }
          },
          "404": {
            "description": "API key not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicError"
                }
              }
            }
          },
          "503": {
            "description": "Auth or API key service is not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "partnerApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Partner API key. The key must be active and include one of the documented apartment read scopes for the endpoint being called."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "adminSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "hope_admin_session",
        "description": "Admin session cookie issued by `/admin/login`. In the internal Swagger UI this cookie is sent automatically when the browser already has an active admin session."
      }
    },
    "schemas": {
      "PublicError": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "error"
            ]
          },
          "code": {
            "type": "string",
            "example": "BAD_REQUEST"
          },
          "message": {
            "type": "string",
            "example": "Invalid API key scopes."
          }
        },
        "required": [
          "status",
          "code",
          "message"
        ]
      },
...
      "ApiKeyScope": {
        "type": "string",
        "enum": [
          "apartments:read",
          "apartments:rental:read",
          "apartments:sale:read"
        ]
      },
      "ApiKeyScopes": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/ApiKeyScope"
        },
        "minItems": 1,
        "uniqueItems": true
      },
      "PartnerApartmentAccessPolicy": {
        "type": "object",
        "properties": {
          "fields": {
            "type": "array",
            "items": {
              "type": "string",
              "example": "rent.warmRent"
            },
            "uniqueItems": true
          }
        },
        "required": [
          "fields"
        ]
      },
      "PartnerAccessPolicy": {
        "type": "object",
        "properties": {
          "apartments": {
            "$ref": "#/components/schemas/PartnerApartmentAccessPolicy"
          }
        }
      },
      "ApiKey": {
        "type": "object",
        "properties": {
          "publicId": {
            "type": "string",
            "example": "hop_live_abc123def456"
          },
          "ownerUserId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "partnerId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "environment": {
            "type": "string",
            "example": "live"
          },
          "keyPrefix": {
            "type": "string",
            "example": "hop_live_abc123def456"
          },
          "role": {
            "type": "string",
            "example": "client"
          },
          "scopes": {
            "$ref": "#/components/schemas/ApiKeyScopes"
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "accessPolicy": {
            "$ref": "#/components/schemas/PartnerAccessPolicy"
          },
          "isActive": {
            "type": "boolean"
          },
          "lastUsedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "revokedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "required": [
          "publicId",
          "ownerUserId",
          "partnerId",
          "name",
          "environment",
          "keyPrefix",
          "role",
          "scopes",
          "notes",
          "accessPolicy",
          "isActive",
          "lastUsedAt",
          "expiresAt",
          "revokedAt",
          "createdAt"
        ]
      },
      "ApiKeyListResponse": {
        "type": "object",
        "properties": {
          "apiKeys": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiKey"
            }
          }
        },
        "required": [
          "apiKeys"
        ]
      },
      "ApiKeyResponse": {
        "type": "object",
        "properties": {
          "apiKey": {
            "$ref": "#/components/schemas/ApiKey"
          }
        },
        "required": [
          "apiKey"
        ]
      },
      "ApiKeyCreateRequest": {
        "type": "object",
        "properties": {
          "partnerId": {
            "type": "string",
            "minLength": 1
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "environment": {
            "type": "string",
            "nullable": true,
            "description": "Optional. Unsupported values normalize to `live` in the service layer."
          },
          "role": {
            "type": "string",
            "nullable": true
          },
          "scopes": {
            "$ref": "#/components/schemas/ApiKeyScopes"
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "accessPolicy": {
            "$ref": "#/components/schemas/PartnerAccessPolicy"
          },
          "expiresAt": {
            "type": "string",
            "nullable": true,
            "description": "Optional expiration timestamp string. The backend does not currently enforce a strict date format."
          }
        },
        "required": [
          "partnerId",
          "name",
          "scopes"
        ]
      },
      "ApiKeyUpdateRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "role": {
            "type": "string",
            "nullable": true
          },
          "scopes": {
            "$ref": "#/components/schemas/ApiKeyScopes"
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "accessPolicy": {
            "$ref": "#/components/schemas/PartnerAccessPolicy"
          },
          "expiresAt": {
            "type": "string",
            "nullable": true,
            "description": "Optional expiration timestamp string. The backend does not currently enforce a strict date format."
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
      "ApiKeyCreateResponse": {
        "type": "object",
        "properties": {
          "apiKey": {
            "$ref": "#/components/schemas/ApiKey"
          },
          "secret": {
            "type": "string",
            "description": "Returned once at creation time."
          }
        },
        "required": [
          "apiKey",
          "secret"
        ]
      },
      "ApiKeyRotateResponse": {
        "type": "object",
        "properties": {
          "previousApiKeyId": {
            "type": "string"
          },
          "apiKey": {
            "$ref": "#/components/schemas/ApiKey"
          },
          "secret": {
            "type": "string",
            "description": "Returned once at rotation time."
          }
        },
        "required": [
          "previousApiKeyId",
          "apiKey",
          "secret"
        ]
      },
      "ApiKeyStats": {
        "type": "object",
        "properties": {
          "totalKeys": {
            "type": "integer"
          },
          "activeKeys": {
            "type": "integer"
          },
          "revokedKeys": {
            "type": "integer"
          },
          "expiredKeys": {
            "type": "integer"
          },
          "lastKeyUseAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "apiKeyUsed24h": {
            "type": "integer"
          },
          "apiKeyAuthFailed24h": {
            "type": "integer"
          }
        },
        "required": [
          "totalKeys",
          "activeKeys",
          "revokedKeys",
          "expiredKeys",
          "lastKeyUseAt",
          "apiKeyUsed24h",
          "apiKeyAuthFailed24h"
        ]
      },
      "ApiKeyStatsResponse": {
        "type": "object",
        "properties": {
          "stats": {
            "$ref": "#/components/schemas/ApiKeyStats"
          }
        },
        "required": [
          "stats"
        ]
      },
...
    }
  }
}